+++ ZooManagement/IDataSource.cs
@@ -8,6 +8,6 @@ namespace ZooManagement
     {
         Dictionary<string, Species> GetSpecies(); 
         List<Animal> GetAnimals(Dictionary<string, Species> species);
-        Dictionary<string, double> GetPrices();
+        Dictionary<string, decimal> GetPrices();
     }
 }
